move all encoding and decoding to version specific files
